+++ flake.nix
@@ -18,7 +18,7 @@
       url = "github:haskell/haskell-language-server";
       flake = false;
     };
-    cardano-node.url = "github:intersectmbo/cardano-node/9.0.0";
+    cardano-node.url = "github:intersectmbo/cardano-node/9.1.0";
     mithril.url = "github:input-output-hk/mithril/2428.0";
     nix-npm-buildpackage.url = "github:serokell/nix-npm-buildpackage";
   };